% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.btvcmfit}
\alias{as.data.frame.btvcmfit}
\title{Extract samples from the btvcmfit object as data frame.}
\usage{
\method{as.data.frame}{btvcmfit}(x, row.names = NULL, optional = FALSE, parameter_types, ...)
}
\description{
Extract samples from the btvcmfit object as data frame.
}
