% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic.R
\name{assign_deterministic}
\alias{assign_deterministic}
\title{Evaluate Definitions and Assign Values of Deterministic Channels}
\usage{
assign_deterministic(data, cl, idx)
}
\arguments{
\item{data}{[\code{data.table}]\cr Data to assign the values into.}

\item{cl}{[\code{language}]\cr A quoted expression defining the channels.}

\item{idx}{[\code{integer()}]\cr A vector of indices.}
}
\description{
Evaluate Definitions and Assign Values of Deterministic Channels
}
