% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamiteformula.R
\name{auxiliary}
\alias{auxiliary}
\alias{aux}
\title{Prepare a deterministic auxiliary channel}
\usage{
auxiliary(formula)

aux(formula)
}
\arguments{
\item{formula}{[\code{formula}]\cr An \R formula describing how the LHS
variable is defined}
}
\description{
Prepare a deterministic auxiliary channel
}
