% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{becausefamily}
\alias{becausefamily}
\alias{categorical}
\alias{gaussian}
\title{Family Functions for \pkg{because} Models}
\usage{
becausefamily(name, ...)

categorical(...)

gaussian(...)
}
\description{
Family Functions for \pkg{because} Models
}
