% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/becauseformula.R
\name{becauseformula}
\alias{becauseformula}
\title{Model formula for \pkg{because}}
\usage{
becauseformula(
  formula,
  family = c("gaussian", "binomial", "poisson", "categorical"),
  data,
  ...
)
}
\description{
Model formula for \pkg{because}
}
