% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btvcm.R
\name{btvcm}
\alias{btvcm}
\title{Bayesian Time-Varying Coefficients Model}
\usage{
btvcm(formula, data, group, time, ...)
}
\arguments{
\item{formula}{todo}

\item{data}{todo}

\item{group}{todo}

\item{time}{todo}

\item{...}{todo}
}
\description{
Bayesian Time-Varying Coefficients Model
}
