% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btvcmformula.R
\name{btvcmformula}
\alias{btvcmformula}
\alias{obs}
\title{Model formula for \pkg{btvcm}}
\usage{
btvcmformula(formula, family, ...)

obs(formula, family, ...)
}
\description{
Model formula for \pkg{btvcm}
}
