% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.btvcmfit}
\alias{coef.btvcmfit}
\title{Extract regression coefficients of btvcmfit}
\usage{
\method{coef}{btvcmfit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{btvcmfit}.}

\item{...}{Ignored.}
}
\description{
Extract regression coefficients of btvcmfit
}
