% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanblocks.R
\name{create_blocks}
\alias{create_blocks}
\title{Create code blocks for the Stan model}
\usage{
create_blocks(formula, ...)
}
\arguments{
\item{formula}{A \code{dynamiteformula} defining the model}

\item{...}{TODO}
}
\description{
Create code blocks for the Stan model
}
