% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite.R
\name{dynamite_backend}
\alias{dynamite_backend}
\title{Check backend capabilities for \code{dynamite}}
\usage{
dynamite_backend(backend)
}
\arguments{
\item{backedn}{Either \code{"cmdstanr"} or \code{"rstan"}}
}
\description{
Check backend capabilities for \code{dynamite}
}
