% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{dynamitefamily}
\alias{dynamitefamily}
\title{Family Functions for \pkg{dynamite} Models}
\usage{
dynamitefamily(name)
}
\arguments{
\item{\[`character(1)`]\cr}{Name of the family.}

\item{...}{TODO is this needed?}
}
\description{
Family Functions for \pkg{dynamite} Models
}
