% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamitefit.R
\name{dynamitefit}
\alias{dynamitefit}
\title{Fit a Bayesian Time-Varying Coefficients Model}
\usage{
dynamitefit(formula, data, group, time, priors = NULL, debug = NULL, ...)
}
\description{
Fit a Bayesian Time-Varying Coefficients Model
}
