% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.btvcmfit}
\alias{fitted.btvcmfit}
\title{Extract fitted values of btvcmfit}
\usage{
\method{fitted}{btvcmfit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{btvcmfit}.}

\item{...}{Ignored.}
}
\description{
Extract fitted values of btvcmfit
}
