% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite.R
\name{formula.dynamitefit}
\alias{formula.dynamitefit}
\title{Access the Model Formula of a Dynamite Model}
\usage{
\method{formula}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object.}

\item{...}{Not used.}
}
\description{
Returns the model definition as a quoted expression.
}
\examples{
formula(gaussian_example_fit)
}
