% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{data}
\name{gaussian_example_fit}
\alias{gaussian_example_fit}
\title{Model Fit for the Simulated Data of a Gaussian Response}
\format{
A \code{dynamitefit} object.
}
\source{
The data was generated via \code{gaussian_example_fit.R} in
\url{https://github.com/ropensci/dynamite/tree/main/data-raw/}
}
\usage{
gaussian_example_fit
}
\description{
A \code{dynamitefit} object obtained by running \code{dynamite} on the
\code{gaussian_example} dataset as
\preformatted{
set.seed(1)
library(dynamite)
gaussian_example_fit <- dynamite(
  obs(y ~ -1 + z + varying(~ x + lag(y)) + random(~1), family = "gaussian") +
    random_spec() + splines(df = 20),
  data = gaussian_example,
  time = "time",
  group = "id",
  iter = 2000,
  warmup = 1000,
  thin = 10,
  chains = 2,
  cores = 2,
  refresh = 0,
  save_warmup = FALSE,
  pars = c("omega_alpha_1_y", "omega_raw_alpha_y", "nu_raw", "nu", "L",
    "sigma_nu", "a_y"),
  include = FALSE
)
}
Note the very small number of samples due to size restrictions on CRAN.
}
\seealso{
Example models:
\code{\link{categorical_example}},
\code{\link{categorical_example_fit}},
\code{\link{gaussian_example}},
\code{\link{multichannel_example}},
\code{\link{multichannel_example_fit}}
}
\concept{examples}
\keyword{datasets}
