% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{data}
\name{gaussian_simulation_fit}
\alias{gaussian_simulation_fit}
\title{Model Fit for the time-varying example in the \code{dynamite_simulation} Vignette}
\format{
A \code{dynamitefit} object.
}
\source{
The data was generated according to a script in
\url{https://github.com/ropensci/dynamite/blob/main/data-raw/gaussian_simulation_fit.R}
}
\usage{
gaussian_simulation_fit
}
\description{
A \code{dynamitefit} object obtained by running \code{dynamite} with the
\code{"Fixed_param"} algorithm on the specified \code{inits} in the example.
\preformatted{
set.seed(1)
library(dynamite)
gaussian_simulation_fit <- dynamite(
  dformula = f,
  data = d,
  time = "time",
  group = "id",
  chains = 1,
  iter = 1,
  algorithm = "Fixed_param",
  init = list(init),
)
}
}
\seealso{
Example models
\code{\link{categorical_example}},
\code{\link{categorical_example_fit}},
\code{\link{gaussian_example}},
\code{\link{gaussian_example_fit}},
\code{\link{multichannel_example}},
\code{\link{multichannel_example_fit}}
}
\concept{examples}
\keyword{datasets}
