% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_helpers.R
\name{generate_sim_call}
\alias{generate_sim_call}
\title{Generate a Quoted Expression to Evaluate Predictions or Fitted Values}
\usage{
generate_sim_call(
  resp,
  resp_levels,
  family,
  type,
  has_fixed,
  has_varying,
  has_fixed_intercept,
  has_varying_intercept,
  has_random_intercept,
  has_offset
)
}
\arguments{
\item{resp}{[\code{character(1)}]\cr Name of the response.}

\item{resp_levels}{[\code{character()}]\cr Levels of a categorical response.}

\item{family}{[\code{dynamitefamily}]\cr Family of the response.}

\item{type}{[\code{character(1)}]\cr Either \code{"predict"} or \code{"fitted"}.}

\item{has_fixed}{[logical(1)]\cr
Does the channel have time-invariant predictors?}

\item{has_varying}{[logical(1)]\cr
Does the channel have time-varying predictors?}

\item{has_fixed_intercept}{[logical(1)]\cr}

\item{has_varying_intercept}{[logical(1)]\cr}

\item{has_random_intercept}{[logical(1)]\cr}

\item{has_offset}{[logical(1)]\cr}
}
\description{
Generate a Quoted Expression to Evaluate Predictions or Fitted Values
}
