% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utilities.R
\name{get_algorithm}
\alias{get_algorithm}
\alias{get_algorithm.stanfit}
\alias{get_algorithm.CmdStanMCMC}
\alias{get_algorithm.CmdStanMCMC_CSV}
\title{Get the algorithm used in a Stan model fit}
\usage{
get_algorithm(x)

\method{get_algorithm}{stanfit}(x)

\method{get_algorithm}{CmdStanMCMC}(x)

\method{get_algorithm}{CmdStanMCMC_CSV}(x)
}
\arguments{
\item{x}{A \code{stanfit} (from \code{rstan}) or a \code{CmdStanMCMC}
(from \code{cmdstanr}) object.}
}
\description{
Get the algorithm used in a Stan model fit
}
\keyword{internal}
