% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_data}
\alias{get_data}
\alias{get_data.dynamiteformula}
\alias{get_data.dynamitefit}
\title{Extract the Model Data of the \pkg{dynamite} Model}
\usage{
get_data(x, ...)

\method{get_data}{dynamiteformula}(x, data, time, group = NULL, ...)

\method{get_data}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamiteformula} or \code{dynamitefit}]\cr The model formula or an
existing \code{dynamitefit} object. See \code{\link[=dynamiteformula]{dynamiteformula()}} and \code{\link[=dynamite]{dynamite()}}.}

\item{...}{Ignored.}

\item{data}{[\code{data.frame}, \code{tibble::tibble}, or \code{data.table::data.table}]\cr
The data that contains the variables in the model in long format.
Supported column types are \code{integer}, \code{logical}, \code{double}, and
\code{factor}. Columns of type \code{character} will be converted to factors.
Unused factor levels will be dropped. The \code{data} can contain missing
values which will simply be ignored in the estimation in a case-wise
fashion (per time-point and per channel). Input \code{data} is converted to
channel specific matrix representations via \code{\link[stats:model.matrix.lm]{stats::model.matrix.lm()}}.}

\item{time}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
time index of observations. If this variable is a factor, the integer
representation of its levels are used internally for defining the time
indexing.}

\item{group}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
unique groups or \code{NULL} corresponding to a scenario without any groups.
If \code{group} is \code{NULL}, a new column \code{.group} is created with constant
value \code{1L} is created indicating that all observations belong to the same
group. In case of name conflicts with \code{data}, see the \code{group_var} element
of the return object to get the column name of the new variable.}
}
\value{
A \code{list} containing the input data to Stan.
}
\description{
Returns the input data to the Stan model. Mostly useful for debugging.
}
\examples{
data.table::setDTthreads(1) # For CRAN
d <- data.frame(y = rnorm(10), x = 1:10, time = 1:10, id = 1)
str(get_data(obs(y ~ x, family = "gaussian"),
  data = d, time = "time", group = "id"
))

}
\seealso{
Model outputs:
\code{\link[=as.data.frame.dynamitefit]{as.data.frame.dynamitefit()}},
\code{\link[=as.data.table.dynamitefit]{as.data.table.dynamitefit()}},
\code{\link[=as_draws_df.dynamitefit]{as_draws_df.dynamitefit()}},
\code{\link[=coef.dynamitefit]{coef.dynamitefit()}},
\code{\link[=confint.dynamitefit]{confint.dynamitefit()}},
\code{\link[=dynamite]{dynamite()}},
\code{\link[=get_code]{get_code()}},
\code{\link[=get_parameter_dims]{get_parameter_dims()}},
\code{\link[=get_parameter_names]{get_parameter_names()}},
\code{\link[=get_parameter_types]{get_parameter_types()}},
\code{\link[=ndraws.dynamitefit]{ndraws.dynamitefit()}},
\code{\link[=nobs.dynamitefit]{nobs.dynamitefit()}}
}
\concept{output}
