% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utilities.R
\name{get_diagnostics}
\alias{get_diagnostics}
\alias{get_diagnostics.stanfit}
\alias{get_diagnostics.CmdStanMCMC}
\alias{get_diagnostics.CmdStanMCMC_CSV}
\title{Get the diagnostics of a Stan model fit}
\usage{
get_diagnostics(x)

\method{get_diagnostics}{stanfit}(x)

\method{get_diagnostics}{CmdStanMCMC}(x)

\method{get_diagnostics}{CmdStanMCMC_CSV}(x)
}
\arguments{
\item{x}{A \code{stanfit} (from \code{rstan}) or a \code{CmdStanMCMC}
(from \code{cmdstanr}) object.}
}
\description{
Get the diagnostics of a Stan model fit
}
\keyword{internal}
