% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utilities.R
\name{get_draws}
\alias{get_draws}
\alias{get_draws.stanfit}
\alias{get_draws.CmdStanMCMC}
\alias{get_draws.CMdStanMCMC_CSV}
\title{Get the draws of a Stan model fit}
\usage{
get_draws(x, pars)

\method{get_draws}{stanfit}(x, pars)

\method{get_draws}{CmdStanMCMC}(x, pars)

\method{get_draws}{CMdStanMCMC_CSV}(x, pars)
}
\arguments{
\item{x}{A \code{stanfit} (from \code{rstan}) or a \code{CmdStanMCMC}
(from \code{cmdstanr}) object.}

\item{pars}{A \code{character} vector of parameter names.}
}
\description{
Get the draws of a Stan model fit
}
\keyword{internal}
