% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utilities.R
\name{get_elapsed_time}
\alias{get_elapsed_time}
\alias{get_elapsed_time.stanfit}
\alias{get_elapsed_time.CmdStanMCMC}
\alias{get_elapsed_time.CmdStanMCMC_CSV}
\title{Get the elapsed time of a Stan model fit}
\usage{
get_elapsed_time(x)

\method{get_elapsed_time}{stanfit}(x)

\method{get_elapsed_time}{CmdStanMCMC}(x)

\method{get_elapsed_time}{CmdStanMCMC_CSV}(x)
}
\arguments{
\item{x}{A \code{stanfit} (from \code{rstan}) or a \code{CmdStanMCMC}
(from \code{cmdstanr}) object.}
}
\description{
Get the elapsed time of a Stan model fit
}
\keyword{internal}
