% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utilities.R
\name{get_max_treedepth}
\alias{get_max_treedepth}
\alias{get_max_treedepth.stanfit}
\alias{get_max_treedepth.CmdStanMCMC}
\alias{get_max_treedepth.CmdStanMCMC_CSV}
\title{Get the maximum treedepth of chains of a Stan model fit}
\usage{
get_max_treedepth(x)

\method{get_max_treedepth}{stanfit}(x)

\method{get_max_treedepth}{CmdStanMCMC}(x)

\method{get_max_treedepth}{CmdStanMCMC_CSV}(x)
}
\arguments{
\item{x}{A \code{stanfit} (from \code{rstan}) or a \code{CmdStanMCMC}
(from \code{cmdstanr}) object.}
}
\description{
Get the maximum treedepth of chains of a Stan model fit
}
\keyword{internal}
