% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utilities.R
\name{get_model_code}
\alias{get_model_code}
\alias{get_model_code.stanfit}
\alias{get_model_code.CmdStanMCMC}
\alias{get_model_code.CmdStanMCMC_CSV}
\title{Get the model code of a Stan model fit}
\usage{
get_model_code(x)

\method{get_model_code}{stanfit}(x)

\method{get_model_code}{CmdStanMCMC}(x)

\method{get_model_code}{CmdStanMCMC_CSV}(x)
}
\arguments{
\item{x}{A \code{stanfit} (from \code{rstan}) or a \code{CmdStanMCMC}
(from \code{cmdstanr}) object.}
}
\description{
Get the model code of a Stan model fit
}
\keyword{internal}
