% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utilities.R
\name{get_nchains}
\alias{get_nchains}
\alias{get_nchains.stanfit}
\alias{get_nchains.CmdStanMCMC}
\alias{get_nchains.CmdStanMCMC_CSV}
\title{Get the number of chains of a Stan model fit}
\usage{
get_nchains(x)

\method{get_nchains}{stanfit}(x)

\method{get_nchains}{CmdStanMCMC}(x)

\method{get_nchains}{CmdStanMCMC_CSV}(x)
}
\arguments{
\item{x}{A \code{stanfit} (from \code{rstan}) or a \code{CmdStanMCMC}
(from \code{cmdstanr}) object.}
}
\description{
Get the number of chains of a Stan model fit
}
\keyword{internal}
