% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utilities.R
\name{get_ndraws}
\alias{get_ndraws}
\alias{get_ndraws.stanfit}
\alias{get_ndraws.CmdStanMCMC}
\alias{get_ndraws.CmdStanMCMC_CSV}
\title{Get the number of draws of a Stan model fit}
\usage{
get_ndraws(x)

\method{get_ndraws}{stanfit}(x)

\method{get_ndraws}{CmdStanMCMC}(x)

\method{get_ndraws}{CmdStanMCMC_CSV}(x)
}
\arguments{
\item{x}{A \code{stanfit} (from \code{rstan}) or a \code{CmdStanMCMC}
(from \code{cmdstanr}) object.}
}
\description{
Get the number of draws of a Stan model fit
}
\keyword{internal}
