% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utilities.R
\name{get_pars_oi}
\alias{get_pars_oi}
\alias{get_pars_oi.stanfit}
\alias{get_pars_oi.CmdStanMCMC}
\alias{get_pars_oi.CmdStanMCMC_CSV}
\title{Get \code{pars_oi} of a Stan model fit}
\usage{
get_pars_oi(x)

\method{get_pars_oi}{stanfit}(x)

\method{get_pars_oi}{CmdStanMCMC}(x)

\method{get_pars_oi}{CmdStanMCMC_CSV}(x)
}
\arguments{
\item{x}{A \code{stanfit} (from \code{rstan}) or a \code{CmdStanMCMC}
(from \code{cmdstanr}) object.}
}
\description{
Get \code{pars_oi} of a Stan model fit
}
\keyword{internal}
