% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{increment_formula_determnistic}
\alias{increment_formula_determnistic}
\title{Add terms to a formula of a deterministic channel}
\usage{
increment_formula_determnistic(formula, x)
}
\arguments{
\item{formula}{A \code{formula} object.}

\item{x}{A \code{character} vector of terms to add.}
}
\description{
Add terms to a formula of a deterministic channel
}
