% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/becauseformula.R
\name{is.becauseformula}
\alias{is.becauseformula}
\title{Checks if argument is a \code{becauseformula} object}
\usage{
is.becauseformula(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{becauseformula} object
}
