% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btvcmformula.R
\name{is.btvcmformula}
\alias{is.btvcmformula}
\title{Checks if argument is a \code{btvcmformula} object}
\usage{
is.btvcmformula(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{btvcmformula} object
}
