% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamiteformula.R
\name{is.dynamiteformula}
\alias{is.dynamiteformula}
\title{Is the Argument a \code{dynamiteformula} Object}
\usage{
is.dynamiteformula(x)
}
\arguments{
\item{x}{An \R object.}
}
\description{
Is the Argument a \code{dynamiteformula} Object
}
