% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{data}
\name{latent_factor_example}
\alias{latent_factor_example}
\title{Simulated Latent Factor Model Panel Data}
\format{
A data frame with 2000 rows and 3 variables:
\describe{
\item{y}{A continuos variable.}
\item{id}{Variable defining individuals (1 to 100).}
\item{time}{Variable defining the time point of the measurement (1 to 20).}
}
}
\source{
The data was generated according to a script in
\url{https://github.com/ropensci/dynamite/blob/main/data-raw/latent_factor_example.R}
}
\usage{
latent_factor_example
}
\description{
A simulated single-channel data containing multiple individuals whose
trajectories are defined by a latent factor and random intercept terms.
}
\seealso{
Example models
\code{\link{categorical_example_fit}},
\code{\link{categorical_example}},
\code{\link{gaussian_example_fit}},
\code{\link{gaussian_example}},
\code{\link{latent_factor_example_fit}},
\code{\link{multichannel_example_fit}},
\code{\link{multichannel_example}}
}
\concept{examples}
\keyword{datasets}
