% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{data}
\name{multichannel_example}
\alias{multichannel_example}
\title{Simulated Multivariate Panel Data}
\format{
A data frame with 3000 rows and 5 variables:
\describe{
\item{id}{Variable defining individuals (1 to 50).}
\item{time}{Variable defining the time point of the measurement (1 to 20).}
\item{g}{Response variable following gaussian distribution.}
\item{p}{Response variable following Poisson distribution.}
\item{b}{Response variable following Bernoulli distribution.}
}
}
\source{
The data was generated via \code{multichannel_example.R} in
\url{https://github.com/ropensci/dynamite/tree/main/data-raw/}
}
\usage{
multichannel_example
}
\description{
A simulated multichannel data containing multiple individuals with multiple
response variables of different distributions.
}
\seealso{
Example models:
\code{\link{categorical_example}},
\code{\link{categorical_example_fit}},
\code{\link{gaussian_example}},
\code{\link{gaussian_example_fit}},
\code{\link{multichannel_example_fit}}
}
\concept{examples}
\keyword{datasets}
