% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ndraws.dynamitefit}
\alias{ndraws.dynamitefit}
\title{Return the Number of Posterior Draws of the dynamitefit Object}
\usage{
\method{ndraws}{dynamitefit}(x)
}
\description{
Return the Number of Posterior Draws of the dynamitefit Object
}
\examples{
ndraws(gaussian_examplefit)
}
