% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite.R
\name{parse_data}
\alias{parse_data}
\title{Parse data for model fitting}
\usage{
parse_data(data, group_var, time_var)
}
\arguments{
\item{data}{[\code{data.frame}]\cr The data frame containing the variables in
the model.}

\item{group_var}{[\code{character(1)},\code{NULL}]\cr Group variable name or \code{NULL}
if there is only one group}

\item{time_var}{[\code{character(1)}]\cr Time index variable name}
}
\description{
Parse data for model fitting
}
