% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo.R
\name{plot.lfo}
\alias{plot.lfo}
\title{Diagnostic Plot for Pareto k Values from LFO}
\usage{
\method{plot}{lfo}(x, ...)
}
\arguments{
\item{x}{x [\code{lfo}]\cr Output from \code{lfo} function.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Plots Pareto k values per each time point (with one point per group),
together with the horizontal line representing the used threshold.
}
\examples{
\dontrun{
plot(lfo(gaussian_example_fit, L = 20))
}
}
