% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_betas}
\alias{plot_betas}
\title{Visualize Time-invariant Regression Coefficients of the Dynamite Model}
\usage{
plot_betas(model, level = 0.05, include_alpha = TRUE)
}
\arguments{
\item{model}{An object of class \code{dynamitefit}.}

\item{level}{[\code{numeric(1)}]\cr Level for posterior intervals.
Default is 0.05, leading to 90\% intervals.}

\item{include_alpha}{[\code{logical(1)}]\cr If \code{TRUE} (default), extracts also
time-invariant intercept term alpha if time-invariant parameters beta are
extracted, and time-varying alpha if time-varying delta are extracted.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize Time-invariant Regression Coefficients of the Dynamite Model
}
\examples{
plot_betas(gaussian_example_fit)
}
