% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_deltas}
\alias{plot_deltas}
\title{Visualize Time-varying Regression Coefficients of the Dynamite Model}
\usage{
plot_deltas(
  model,
  level = 0.05,
  alpha = 0.5,
  scales = c("fixed", "free"),
  include_alpha = TRUE
)
}
\arguments{
\item{model}{[\code{dynamitefit}]\cr The model fit object}

\item{level}{[\code{numeric(1)}]\cr Level for posterior intervals.
Default is 0.05, leading to 90\% intervals.}

\item{alpha}{[\code{numeric(1)}]\cr Opacity level for \code{geom_ribbon}.
Default is 0.5.}

\item{scales}{[\code{character(1)}] Should y-axis of the panels be \code{"fixed"}
(the default) or \code{"free"}? See \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{include_alpha}{[\code{logical(1)}]\cr If \code{TRUE} (default), extracts also
time-invariant intercept term alpha if time-invariant parameters beta are
extracted, and time-varying alpha if time-varying delta are extracted.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize Time-varying Regression Coefficients of the Dynamite Model
}
\examples{
plot_deltas(gaussian_example_fit, scales = "free") +
  ggplot2::theme_minimal()

}
