% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamiteformula.R
\name{+.dynamiteformula}
\alias{+.dynamiteformula}
\title{Join two dynamiteformulas}
\usage{
\method{+}{dynamiteformula}(e1, e2)
}
\arguments{
\item{e1}{An \R object}

\item{e2}{An \R object}
}
\description{
Join two dynamiteformulas
}
