% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.btvcmfit}
\alias{print.btvcmfit}
\title{Print Summary of btvcm Object}
\usage{
\method{print}{btvcmfit}(x, pars, ...)
}
\arguments{
\item{x}{An output from \code{\link{btvcm}}.}

\item{pars}{A character vector of parameter names.
The default is to print summary of spline coefficient standard deviations tau.}

\item{...}{Additional arguments to \code{\link{print.stanfit}}.}
}
\description{
Prints the summary information of time-invariant model parameters.
}
