% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dynamitefit}
\alias{print.dynamitefit}
\title{Print Summary of Dynamite Object}
\usage{
\method{print}{dynamitefit}(x, hmc_diagnostics = TRUE, ...)
}
\arguments{
\item{x}{An output from \code{\link{dynamite}}.}

\item{hmc_diagnostics}{[\code{logical(1)}]\cr if \code{TRUE} (default), prints the
summary of Stan's HMC sampler diagnostics. See \code{\link[rstan:check_hmc_diagnostics]{rstan::check_hmc_diagnostics()}}
for details.}

\item{...}{Ignored.}
}
\description{
Prints the summary information of the estimated dynamite model.
}
