% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{states}
\alias{states}
\title{Define the hidden states of the model.}
\usage{
states(n, fixed = c(1L, 1L), individual = FALSE, replace = FALSE)
}
\arguments{
\item{n}{An integer value giving the number of hidden states.}

\item{fixed}{An integer vector with two elements giving
the number of hidden states to fix at the head and tail.}

\item{individual}{A logical value. If \code{FALSE} (the default), then
the subjects are governed by the same hidden state process. If \code{TRUE},
each subject will have an individual process.}

\item{override}{A logical value. If \code{FALSE} (the default), an existing
definition for the hidden states will not be overridden by another call to \code{states}.
If \code{TRUE}, any existing definitions will be replaced.}
}
\description{
Define the hidden states of the model.
}
