% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dynamitefit}
\alias{summary.dynamitefit}
\title{Summary for a Dynamite Model Fit}
\usage{
\method{summary}{dynamitefit}(object, ...)
}
\arguments{
\item{object}{[\code{dynamitefit}]\cr The model fit object.}

\item{...}{Further arguments to the \code{as.data.frame.dynamitefit}.}
}
\description{
Returns the summary statistics of the posterior samples of the model; this
is an alias of \code{as.data.frame(object)}.
}
\examples{
summary(gaussian_example_fit, types = "beta")

}
