% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldata.R
\name{use_data}
\alias{use_data}
\title{Define the data used by the model.}
\usage{
use_data(data, replace = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} in long format,}

\item{override}{A logical value. If \code{FALSE} (the default), an existing
definition for the model data will not be overridden by another call to \code{use_data}.
If \code{TRUE}, any existing definitions will be replaced.}
}
\description{
Define the data used by the model.
}
