% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.dynamitefit}
\alias{as.data.frame.dynamitefit}
\title{Extract Samples From the dynamitefit Object as a Data Frame.}
\usage{
\method{as.data.frame}{dynamitefit}(
  x,
  row.names = NULL,
  optional = FALSE,
  responses = NULL,
  types = NULL,
  ...
)
}
\arguments{
\item{x}{The estimated \code{dynamite} model.}

\item{row.names}{\code{NULL} (default) or a character vector giving the row
names for the data frame.}

\item{optional}{Ignored.}

\item{responses}{TODO}

\item{types}{TODO}

\item{...}{Ignored.}
}
\description{
Extract Samples From the dynamitefit Object as a Data Frame.
}
