% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.dynamitefit}
\alias{as.data.frame.dynamitefit}
\title{Extract samples from the dynamitefit object as data frame.}
\usage{
\method{as.data.frame}{dynamitefit}(x, row.names = NULL, optional = FALSE, parameter_types, ...)
}
\arguments{
\item{x}{The estimated \code{dynamite} model.}

\item{row.names}{\code{NULL} (default) or a character vector giving the row
names for the data frame.}

\item{optional}{Ignored.}

\item{parameter_types}{What type of parameters should be returned? Possible
choices are \code{"beta"}, \code{"tau"}, \code{"a"}, \code{"beta"}, or
\code{"all"} (default). #TODO sigmas etc? others? Or just return all of these...}

\item{...}{Ignored.}
}
\description{
Extract samples from the dynamitefit object as data frame.
}
