% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.dynamitefit}
\alias{as.data.frame.dynamitefit}
\title{Extract Samples From the dynamitefit Object as a Data Frame.}
\usage{
\method{as.data.frame}{dynamitefit}(
  x,
  row.names = NULL,
  optional = FALSE,
  responses = NULL,
  types = NULL,
  ...
)
}
\arguments{
\item{x}{The estimated \code{dynamite} model.}

\item{row.names}{Ignored.}

\item{optional}{Ignored.}

\item{responses}{[\code{character()}]\cr Response(s) for which the  samples
should be extracted.}

\item{types}{[\code{character()}]\cr Type(s) of the parameters for which the
samples should be extracted.}

\item{...}{Ignored.}
}
\description{
You can use the arguments \code{responses} and \code{types} to extract only a subset
of the model parameters (i.e. only certain types of parameters related to a
certain response variable). The prior data frame (from \code{get_priors}) shows
potential values for these variables.
}
\note{
The spline coefficients \code{alpha} are never returned, but they can be
obtained with \code{as_draws} function among all the other parameters.
}
\examples{
fit <- dynamite(
  obs(y ~ -1 + x + varying(~1), family = gaussian()) + splines(df = 10),
  data = gaussian_example, time = time, group = id,
  iter = 100, chains = 1, refresh = 0)
}
