% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws.R
\name{as_draws_df.dynamitefit}
\alias{as_draws_df.dynamitefit}
\alias{as_draws}
\alias{as_draws_df}
\alias{as_draws.dynamitefit}
\title{Convert \code{dynamite} Output to \code{draws_df} Format}
\usage{
\method{as_draws_df}{dynamitefit}(x, parameter_types, ...)

\method{as_draws}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dynamitefit}.}

\item{parameter_types}{What parameters should be returned? Possible choices are
\code{"beta"}, \code{"tau"}, ... #TODO}

\item{...}{Ignored.}
}
\value{
A \code{draws_df} object.
}
\description{
Converts the output from \code{dynamite} call to a
\code{draws_df} format of the \code{posterior} package, enabling the use
of diagnostics and plotting methods of \code{posterior} and \code{bayesplot}
packages.
}
