% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws.R
\name{as_draws_df.dynamitefit}
\alias{as_draws_df.dynamitefit}
\alias{as_draws}
\alias{as_draws_df}
\alias{as_draws.dynamitefit}
\title{Convert \code{dynamite} Output to \code{draws_df} Format}
\usage{
\method{as_draws_df}{dynamitefit}(x, responses = NULL, types = NULL, ...)

\method{as_draws}{dynamitefit}(x, responses = NULL, types = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{dynamitefit}.}

\item{responses}{[\code{character()}]\cr Response(s) for which the  samples
should be extracted. Possible options are \code{unique(x$priors$response)}}

\item{types}{[\code{character()}]\cr Type(s) of the parameters for which the
samples should be extracted. Possible options are \code{unique(x$priors$type)}.}

\item{...}{Ignored.}
}
\value{
A \code{draws_df} object.
}
\description{
Converts the output from \code{dynamite} call to a
\code{draws_df} format of the \code{posterior} package, enabling the use
of diagnostics and plotting methods of \code{posterior} and \code{bayesplot}
packages. Note that this function returns all variables in a wide format,
whereas \code{\link[=as.data.frame]{as.data.frame()}} uses the long format.
}
\examples{
as_draws(gaussian_example_fit, types = c("sigma", "beta"))

}
