% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/becauseformula.R
\name{becauseformula}
\alias{becauseformula}
\alias{obs}
\title{Model formula for \pkg{because}}
\usage{
becauseformula(formula, family, ...)

obs(formula, family, ...)
}
\description{
Model formula for \pkg{because}
}
