% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{btvcmfamily}
\alias{btvcmfamily}
\alias{categorical}
\alias{gaussian}
\title{Family Functions for \pkg{btvcm} Models}
\usage{
btvcmfamily(name, ...)

categorical(...)

gaussian(...)
}
\description{
Family Functions for \pkg{btvcm} Models
}
