% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{data}
\name{categorical_example}
\alias{categorical_example}
\title{Simulated Categorical Multivariate Panel Data}
\format{
A data frame with 2000 rows and 5 variables:
\describe{
\item{id}{Variable defining individuals (1 to 100).}
\item{time}{Variable defining the time point of the measurement (1 to 20).}
\item{x}{Categorical variable with three levels, A, B, and C.}
\item{y}{Categorical variable with three levels, a, b, and c.}
\item{z}{A continuous covariate.}
}
}
\source{
The data was generated via \code{categorical_example.R} in
\url{https://github.com/ropensci/dynamite/tree/main/data-raw/}
}
\usage{
categorical_example
}
\description{
A simulated data containing multiple individuals with two categorical
response variables.
}
\seealso{
Example models
\code{\link{categorical_example_fit}},
\code{\link{gaussian_example}},
\code{\link{gaussian_example_fit}},
\code{\link{multichannel_example}},
\code{\link{multichannel_example_fit}}
}
\concept{examples}
\keyword{datasets}
