% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{data}
\name{categorical_example_fit}
\alias{categorical_example_fit}
\title{Model Fit for the categorical_example Data}
\format{
A \code{dynamitefit} object.
}
\source{
Script in
\url{https://github.com/santikka/dynamite/blob/main/data-raw/categorical_example_fit.R}
}
\usage{
categorical_example_fit
}
\description{

}
\section{note the small number of samples due to size restrictions on CRAN}{

}

\keyword{datasets}
