% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{data}
\name{categorical_example_fit}
\alias{categorical_example_fit}
\title{Model Fit for the categorical_example Data}
\format{
A \code{dynamitefit} object.
}
\source{
The data was generated according to a script in
\url{https://github.com/santikka/dynamite/blob/main/data-raw/categorical_example_fit.R}
}
\usage{
categorical_example_fit
}
\description{
TODO do we want to use the term "channel"??
}
\details{
A \code{dynamitefit} object obtained by running a \code{dynamite} on the
\code{categorical_example} dataset as
\preformatted{
set.seed(1)
library(dynamite)
f <- obs(g ~ -1 + lag(g) + lag(logp) + varying(~x), family = gaussian()) +
  obs(p ~ x + lag(g) + lag(logp) + lag(b), family = poisson()) +
  obs(b ~ lag(b) * lag(logp) + lag(b) * x + lag(b) * lag(g),
      family = bernoulli()) +
  aux(numeric(logp) ~ log(p + 1)) + splines(df = D)


categorical_example_fit <- dynamite(f, categorical_example, "id", "time",
 chains = 2, cores = 2, init = 0, refresh = 0)
}
}
\keyword{datasets}
