% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.dynamitefit}
\alias{coef.dynamitefit}
\title{Extract regression coefficients of dynamitefit}
\usage{
\method{coef}{dynamitefit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{dynamitefit}.}

\item{...}{Ignored.}
}
\description{
Extract regression coefficients of dynamitefit
}
