% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dynamite-deprecated}
\alias{dynamite-deprecated}
\alias{plot_betas}
\alias{plot_deltas}
\alias{plot_nus}
\alias{plot_lambdas}
\alias{plot_psis}
\title{Deprecated Functions in the dynamite Package}
\usage{
plot_betas(x, ...)
plot_deltas(x,  ...)
plot_nus(x, ...)
plot_lambdas(x,  ...)
plot_psis(x, ...)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object.}

\item{...}{Not used.}
}
\value{
A \code{ggplot} object.
}
\description{
These functions are provided for compatibility with older versions of the
package. They will eventually be completely removed.
}
\section{Details}{

\itemize{
\item \code{plot_betas} is now called via \code{plot(., plot_type = "beta")}
\item \code{plot_deltas} is now called via \code{plot(., plot_type = "delta")}
\item \code{plot_nus} is now called via \code{plot(., plot_type = "nu")}
\item \code{plot_lambdas} is now called via \code{plot(., plot_type = "lambda")}
\item \code{plot_psis} is now called via \code{plot(., plot_type = "psi")}
}
}

\seealso{
See \code{\link[=plot.dynamitefit]{plot.dynamitefit()}} for documentation of the
parameters of these functions
}
