% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{package}
\name{dynamite-package}
\alias{dynamite-package}
\title{The 'dynamite' package.}
\description{
Easy-to-use and efficient interface for Bayesian inference of
complex panel data consisting of multiple individuals with multiple
measurements over time. Supports several observational distributions,
time-varying effects and realistic counterfactual predictions which take into
account the dynamic structure of the model. See the package vignette for
details.
}
