% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite.R
\name{dynamite}
\alias{dynamite}
\title{Estimate a Bayesian Dynamic Multivariate Panel Model}
\usage{
dynamite(formula, data, group, time, priors = NULL, debug = NULL, ...)
}
\arguments{
\item{formula}{todo}

\item{data}{todo}

\item{group}{todo}

\item{time}{todo}

\item{...}{todo}
}
\description{
Estimate a Bayesian Dynamic Multivariate Panel Model
}
