% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite.R
\name{dynamite}
\alias{dynamite}
\title{Estimate a Bayesian Dynamic Multivariate Panel Model}
\usage{
dynamite(dformula, data, group, time, priors = NULL, debug = NULL, ...)
}
\arguments{
\item{dformula}{[\code{dynamiteformula}]\cr The model formula. See 'Details'.}

\item{data}{[\code{data.frame}]\cr The data frame containing the variables in
the model.
TODO document here allowed data types etc.}

\item{group}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
unique groups.}

\item{time}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
time axis.}

\item{priors}{TODO}

\item{debug}{TODO}

\item{...}{Additional arguments to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}.}
}
\description{
Estimate a Bayesian Dynamic Multivariate Panel Model
}
\examples{
\dontrun{
fit <- dynamite(obs(y ~ -1 + varying(~x), family = gaussian()) +
  lags("varying") + splines(df = 20), gaussian_example, id, time,
  chains = 1, refresh = 0)

library(dplyr)
library(ggplot2)
cf <- coef(fit) \%>\%
  group_by(time, variable) \%>\%
  summarise(
    mean = mean(value),
    lwr = quantile(value, 0.025),
    upr = quantile(value, 0.975))

cf \%>\%
  ggplot(aes(time, mean)) + theme_bw() +
    geom_ribbon(aes(ymin = lwr, ymax = upr), alpha = 0.7) +
    geom_line() +
    facet_wrap(~ variable, scales = "free_y")
}
}
