% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite.R
\name{dynamite}
\alias{dynamite}
\title{Estimate a Bayesian Dynamic Multivariate Panel Model}
\usage{
dynamite(formula, data, group, time, priors = NULL, debug = NULL, ...)
}
\arguments{
\item{formula}{[\code{dynamiteformula}]\cr The model formula. See 'Details'.}

\item{data}{[\code{data.frame}]\cr The data frame containing the variables in
the model.}

\item{group}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
unique groups.}

\item{time}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
time axis.}

\item{priors}{TODO}

\item{debug}{TODO}

\item{...}{TODO}
}
\description{
Estimate a Bayesian Dynamic Multivariate Panel Model
}
