% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamiteformula.R
\name{dynamiteformula}
\alias{dynamiteformula}
\alias{obs}
\title{Model formula for \pkg{dynamite}}
\usage{
dynamiteformula(formula, family, ...)

obs(formula, family, ...)
}
\description{
Model formula for \pkg{dynamite}
}
