% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamiteformula.R
\name{dynamiteformula}
\alias{dynamiteformula}
\alias{obs}
\alias{aux}
\title{Model formula for \pkg{dynamite}}
\usage{
dynamiteformula(formula, family, random_intercept = FALSE)

obs(formula, family, random_intercept = FALSE)

aux(formula)
}
\arguments{
\item{formula}{[\code{formula}]\cr An R formula describing the model.}

\item{family}{[\code{character(1)}]\cr The family name.}

\item{random_intercept}{[\code{logical(1)}]\cr If \code{TRUE}, adds
individual-level intercepts to the channel.}
}
\description{
TODO description and details of all features
TODO document families here as well
TODO document lag-conversion to data
}
\section{Functions}{
\itemize{
\item \code{aux}: Prepare a Deterministic Auxiliary Channel
}}

