% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamiteformula.R
\name{dynamiteformula}
\alias{dynamiteformula}
\alias{obs}
\title{Model formula for \pkg{dynamite}}
\usage{
dynamiteformula(formula, family, ...)

obs(formula, family, ...)
}
\arguments{
\item{formula}{[\code{formula}]\cr An R formula describing the model.}

\item{family}{[\code{call}]\cr A call to a family function, e.g., \code{gaussian()}.}

\item{...}{TODO}
}
\description{
TODO description and details of all features
}
