% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.btvcmfit}
\alias{fitted.btvcmfit}
\title{Extract fitted values of btvcmfit}
\usage{
\method{fitted}{btvcmfit}(object, newdata = NULL, n_draws = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{btvcmfit}.}

\item{...}{Ignored.}
}
\description{
TODO Note that these are conditional on the observed data i.e., we don't simulate new lagged values for covariates
}
