% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.dynamitefit}
\alias{fitted.dynamitefit}
\title{Extract Fitted Values of a Dynamite Model}
\usage{
\method{fitted}{dynamitefit}(object, n_draws = NULL, ...)
}
\arguments{
\item{object}{[\code{dynamitefit}]\cr The model fit object.}

\item{n_draws}{[\code{integer(1)}]\cr Number of posterior samples to use,
default is \code{NULL} which uses all samples.}

\item{...}{Ignored.}
}
\description{
Note that these are conditional on the observed data i.e., we don't
simulate new lagged values for covariates, so we underestimate the
uncertainty. It is typically better to use predict with \code{type = "mean"}.
These fitted value are mostly useful only for studying one-step ahead
estimates.
}
\examples{
fitted(gaussian_example_fit, n_draws = 2L)

}
