% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.dynamitefit}
\alias{fitted.dynamitefit}
\title{Extract fitted values of dynamitefit}
\usage{
\method{fitted}{dynamitefit}(object, newdata = NULL, n_draws = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{dynamitefit}.}

\item{...}{Ignored.}
}
\description{
TODO Note that these are conditional on the observed data i.e., we don't
simulate new lagged values for covariates, so we underestimate the uncertainty.
Better to use predict with type = "mean" TODO do it
}
