% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite.R
\name{formula.dynamitefit}
\alias{formula.dynamitefit}
\title{Access The Model Formula of a \code{dynamitefit} Object}
\usage{
\method{formula}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object.}

\item{...}{Not used.}
}
\description{
Returns the model definition as a quoted expression.
}
