% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite.R
\name{formula.dynamitefit}
\alias{formula.dynamitefit}
\title{Access model formula of a dynamite model fit}
\usage{
\method{formula}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object}
}
\description{
Return a list containing the formulas defining each channel of the model.
}
