% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{data}
\name{gaussian_example}
\alias{gaussian_example}
\title{Simulated Data of Gaussian Responses}
\format{
A data frame with 3000 rows and 4 variables:
\describe{
\item{y}{The response variable}
\item{x}{Single continuous covariate}
\item{id}{Variable defining individuals (1 to 100)}
\item{time}{Variable defining the time point of the measurement (1 to 30)}
}
}
\usage{
gaussian_example
}
\description{
A simulated data containing gaussian response variables with one covariate.
The data was generated from a model with intercept, effect of
covariate x and the lagged value of the response variable, with all effects
being time-varying according to a spline with 20 degrees of freedom.
}
\keyword{datasets}
