% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{data}
\name{gaussian_example_fit}
\alias{gaussian_example_fit}
\title{Model Fit for the Simulated Data of Gaussian Responses}
\format{
A \code{dynamitefit} object.
}
\source{
The data was generated according to a script in
\url{https://github.com/santikka/dynamite/blob/main/data-raw/gaussian_example_fit.R}
}
\usage{
gaussian_example_fit
}
\description{

}
\details{

}
\keyword{datasets}
