% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_priors.R
\name{get_code}
\alias{get_code}
\title{TODO documentation}
\usage{
get_code(dformula, data, group, time)
}
\arguments{
\item{dformula}{[\code{dynamiteformula}]\cr The model formula.
See \code{\link[=dynamiteformula]{dynamiteformula()}}.}

\item{data}{[\code{data.frame}]\cr The data frame containing the variables in
the model.}

\item{group}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
unique groups.}

\item{time}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
time axis.}
}
\description{
TODO documentation
}
