% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_gets.R
\name{get_code}
\alias{get_code}
\alias{get_code.dynamiteformula}
\alias{get_code.dynamitefit}
\title{Extract the Stan Code of the Dynamite Model}
\usage{
get_code(x, ...)

\method{get_code}{dynamiteformula}(x, data, group, time, ...)

\method{get_code}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamiteformula} or \code{dynamitefit}]\cr The model formula or an
existing \code{dynamitefit} object. See \code{\link[=dynamiteformula]{dynamiteformula()}} and \code{\link[=dynamite]{dynamite()}}.}

\item{...}{Ignored.}

\item{data}{[\code{data.frame}]\cr The data frame containing the variables in
the model.}

\item{group}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
unique groups.}

\item{time}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
time axis.}
}
\description{
Returns the Stan code of the model. Mostly useful for debugging or for
building a customized version of the model.
}
\examples{

d <- data.frame(y = rnorm(10), x = 1:10, time = 1:10, id = 1)
cat(get_code(obs(y ~ x, family = "gaussian"),
  data = d, time = "time", group = "id"))
}
