% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_priors.R
\name{get_code}
\alias{get_code}
\alias{get_code.dynamitefit}
\title{Extract the Stan Code of the Dynamite Model}
\usage{
get_code(x, data, group, time, ...)

\method{get_code}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamiteformula} or \code{dynamitefit}]\cr The model formula or
existing \code{dynamitefit} object. See \code{\link[=dynamiteformula]{dynamiteformula()}} and \code{\link[=dynamite]{dynamite()}}.}

\item{data}{[\code{data.frame}]\cr The data frame containing the variables in
the model.}

\item{group}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
unique groups.}

\item{time}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
time axis.}

\item{...}{Ignored.}
}
\description{
Returns the Stan code of the model. Mostly useful for debugging or for
building customized version of the model.
}
