% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_gets.R
\name{get_parameter_dims}
\alias{get_parameter_dims}
\alias{get_parameter_dims.dynamiteformula}
\alias{get_parameter_dims.dynamitefit}
\title{Get Parameter Dimensions of the Dynamite Model}
\usage{
get_parameter_dims(x, ...)

\method{get_parameter_dims}{dynamiteformula}(x, data, time, group = NULL, ...)

\method{get_parameter_dims}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object.}

\item{...}{Ignored}
}
\value{
A named list with all parameter dimensions of the input model.
}
\description{
Extracts the names and dimensions of all parameters used in the
\code{dynamitefit} object. See also \code{\link[=get_parameter_types]{get_parameter_types()}} and
\code{\link[=get_parameter_names]{get_parameter_names()}}. The returned dimensions match those of
the \code{stanfit} element of the \code{dynamitefit} object.
}
\examples{
get_parameter_dims(multichannel_example_fit)
}
\seealso{
Model outputs
\code{\link{as.data.frame.dynamitefit}()},
\code{\link{as.data.table.dynamitefit}()},
\code{\link{as_draws_df.dynamitefit}()},
\code{\link{coef.dynamitefit}()},
\code{\link{confint.dynamitefit}()},
\code{\link{dynamite}()},
\code{\link{get_code}()},
\code{\link{get_data}()},
\code{\link{get_parameter_names}()},
\code{\link{get_parameter_types}()},
\code{\link{ndraws.dynamitefit}()},
\code{\link{nobs.dynamitefit}()}
}
\concept{output}
