% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_gets.R
\name{get_parameter_types}
\alias{get_parameter_types}
\alias{get_parameter_types.dynamitefit}
\title{Get Parameter Types of the Dynamite Model}
\usage{
get_parameter_types(x, ...)

\method{get_parameter_types}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr A dynamitefit object.}

\item{...}{Ignored.}
}
\value{
A character vector with all parameter types of the input model.
}
\description{
Extracts all parameter types of used in the \code{dynamitefit} object. See
\code{\link[=as.data.frame.dynamitefit]{as.data.frame.dynamitefit()}} for explanations of different types.
}
\examples{
get_parameter_types(multichannel_example_fit)
}
