% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_gets.R
\name{get_priors}
\alias{get_priors}
\alias{get_priors.dynamiteformula}
\alias{get_priors.dynamitefit}
\title{Get Prior Definitions of a Dynamite Model}
\usage{
get_priors(x, ...)

\method{get_priors}{dynamiteformula}(x, data, group, time, ...)

\method{get_priors}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamiteformula} or \code{dynamitefit}]\cr The model formula or an
existing \code{dynamitefit} object. See \code{\link[=dynamiteformula]{dynamiteformula()}} and \code{\link[=dynamite]{dynamite()}}.}

\item{...}{Ignored.}

\item{data}{[\code{data.frame}]\cr The data frame containing the variables in
the model.}

\item{group}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
unique groups.}

\item{time}{[\code{character(1)}]\cr A column name of \code{data} that denotes the
time axis.}
}
\description{
Extracts the priors used in the dynamite model as a data frame. You
can then alter the priors by changing the contents of the \code{prior} column and
supplying this data frame to \code{dynamite} function using the argument \code{priors}.
}
\details{
Note that the prior for the intercept term \code{alpha} is actually defined
in a centered form, so the prior is related to the \code{alpha} when the
covariates at the first time point are centered around their means. In other
words, the prior is defined for \code{alpha + x_m * gamma} where \code{x_m} is vector
of covariate means and gamma contains the corresponding coefficients (\code{beta}
and \code{delta_1}). If you want to use prior directly on \code{alpha}, remove
intercept from the formula and add a dummy covariate consisting of ones to
the model.
}
\note{
Only the \code{prior} column of the output should be altered when defining
the user-defined priors for the \code{dynamite}.
}
\examples{

d <- data.frame(y = rnorm(10), x = 1:10, time = 1:10, id = 1)
get_priors(obs(y ~ x, family = "gaussian"),
  data = d, time = "time", group = "id")

}
