% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{hmc_diagnostics}
\alias{hmc_diagnostics}
\alias{hmc_diagnostics.dynamitefit}
\title{HMC Diagnostics for a \pkg{dynamite} Model}
\usage{
hmc_diagnostics(x, ...)

\method{hmc_diagnostics}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} (invisibly).
data.table::setDTthreads(1) # For CRAN
hmc_diagnostics(gaussian_example_fit)
}
\description{
Prints the divergences, saturated treedepths, and low E-BFMI warnings.
}
\seealso{
Model diagnostics
\code{\link{lfo}()},
\code{\link{loo.dynamitefit}()},
\code{\link{mcmc_diagnostics}()}
}
\concept{diagnostics}
