% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{hmc_diagnostics}
\alias{hmc_diagnostics}
\title{HMC Diagnostics for a Dynamite Model}
\usage{
hmc_diagnostics(x)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object.}
}
\description{
Prints the divergences, saturated treedepths, and low E-BFMI warnings.
}
\seealso{
Model diagnostics
\code{\link{lfo}()},
\code{\link{loo.dynamitefit}()},
\code{\link{mcmc_diagnostics}()}
}
\concept{diagnostics}
