% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamiteformula.R
\name{is.dynamiteformula}
\alias{is.dynamiteformula}
\title{Checks if argument is a \code{dynamiteformula} object}
\usage{
is.dynamiteformula(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{dynamiteformula} object
}
