% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lags.R
\name{lags}
\alias{lags}
\title{Add each lagged response as a predictor to each channel.}
\usage{
lags(k = 1L, type = c("fixed", "varying"))
}
\arguments{
\item{k}{[\code{integer(1)}: \sQuote{1}]\cr Indicates how many previous
observations should be included.}

\item{type}{[\code{integer(1)}: \sQuote{"fixed"}]\cr Either
\code{"fixed"} or \code{"varying"} which indicates whether the coefficients of the
lag terms should vary in time or not.}
}
\description{
Add each lagged response as a predictor to each channel.
}
