% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lags.R
\name{lags}
\alias{lags}
\title{Add Each Lagged Response as a Predictor to Each Channel.}
\usage{
lags(k = 1L, type = c("fixed", "varying"))
}
\arguments{
\item{k}{[\code{integer()}: \sQuote{1}]\cr
Lagged values indicated by \code{k}  of each observed response variable
will be added as a predictor for each channel.}

\item{type}{[\code{integer(1)}: \sQuote{"fixed"}]\cr Either
\code{"fixed"} or \code{"varying"} which indicates whether the coefficients of the
added lag terms should vary in time or not.}
}
\description{
Add Each Lagged Response as a Predictor to Each Channel.
}
