% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{lags}
\alias{lags}
\title{Add each lagged response as a predictor to each channel. Overrides all other lag definitions.}
\usage{
lags(k = 1L)
}
\arguments{
\item{k}{An integer indicating how many previous observations should be included. Defaults to 1.}
}
\description{
Add each lagged response as a predictor to each channel. Overrides all other lag definitions.
}
