% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lags.R
\name{lags}
\alias{lags}
\title{Add Lagged Responses as Predictors to Each Channel of a Dynamite Model}
\usage{
lags(k = 1L, type = c("fixed", "varying"))
}
\arguments{
\item{k}{[\code{integer()}: \sQuote{1}]\cr
Values lagged by \code{k} units of time of each observed response variable
will be added as a predictor for each channel.}

\item{type}{[\code{integer(1)}: \sQuote{"fixed"}]\cr Either
\code{"fixed"} or \code{"varying"} which indicates whether the coefficients of the
added lag terms should vary in time or not.}
}
\description{
Add Lagged Responses as Predictors to Each Channel of a Dynamite Model
}
\examples{
obs(y ~ -1 + varying(~x), family = "gaussian") +
  lags(type = "varying") + splines(df = 20)

}
