% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo.R
\name{lfo}
\alias{lfo}
\title{Approximate Leave-Future-Out (LFO) Cross-validation}
\usage{
lfo(x, L, verbose = TRUE, k_threshold = 0.7, ...)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object.}

\item{L}{Positive integer defining how many observations should be used for
the initial fit.}

\item{verbose}{If \code{TRUE} (default), print the progress of the LFO
computations to the console.}

\item{k_threshold}{Threshold for the pareto k estimate triggering refit. Default
is 0.7.}

\item{...}{Additional parameters to \code{dynamite}.}
}
\value{
List with components \code{ELPD} (Expected log predictive density),
\code{ELPDs} (observation-specific ELPDs for each time point),
\code{ks} (Pareto k values), and \code{refits} (time points where model
was re-estimated).
}
\description{
Estimates the leave-future-out (LFO) information criterion for \code{dynamite}
models using Pareto smoothed importance sampling.
}
\examples{
\dontrun{
# this gives warnings due to the small number of iterations
lfo(gaussian_example_fit, L = 20)
}
}
