% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{mcmc_diagnostics}
\alias{mcmc_diagnostics}
\title{Diagnostic Values of Dynamite Model}
\usage{
mcmc_diagnostics(x, n = 1)
}
\arguments{
\item{x}{A \code{dynamitefit}object.}

\item{n}{[\code{integer(1)}]\cr How many rows to print in
parameter-specific convergence measures. Default is 1.}
}
\description{
Prints HMC diagnostics, and lists parameters with smallest effective sample
sizes and largest Rhat values. See \code{\link[rstan:check_hmc_diagnostics]{rstan::check_hmc_diagnostics()}} and
\code{\link[posterior:draws_summary]{posterior::default_convergence_measures()}} for details.
}
