% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamite-package.R
\docType{data}
\name{multichannel_example}
\alias{multichannel_example}
\title{Simulated Multivariate Panel Data}
\format{
A data frame with 3000 rows and 5 variables:
\describe{
\item{y}{The response variable}
\item{x}{Single continuous covariate}
\item{z}{Single binary covariate}
\item{id}{Variable defining individuals (1 to 100)}
\item{time}{Variable defining the time point of the measurement (1 to 30)}
}
}
\source{
The data was generated according to a script in
\url{https://github.com/santikka/dynamite/blob/main/data-raw/categorical_example.R}
}
\usage{
multichannel_example
}
\description{
A simulated data containing multiple individuals with multiple response
variables.
}
\keyword{datasets}
