% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ndraws.dynamitefit}
\alias{ndraws.dynamitefit}
\alias{ndraws}
\title{Return the Number of Posterior Draws of the dynamitefit Object}
\usage{
\method{ndraws}{dynamitefit}(x)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object.}
}
\description{
Return the Number of Posterior Draws of the dynamitefit Object
}
\examples{
ndraws(gaussian_example_fit)
}
