% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs.dynamitefit}
\alias{nobs.dynamitefit}
\title{Extract the Number of Observations Used to Fit a \pkg{dynamite} Model}
\usage{
\method{nobs}{dynamitefit}(object, ...)
}
\arguments{
\item{object}{[\code{dynamitefit}]\cr The model fit object.}

\item{...}{Not used.}
}
\value{
Total number of non-missing observations as an \code{integer}.
}
\description{
Extract the Number of Observations Used to Fit a \pkg{dynamite} Model
}
\examples{
data.table::setDTthreads(1) # For CRAN
nobs(gaussian_example_fit)

}
\seealso{
Model outputs
\code{\link{as.data.frame.dynamitefit}()},
\code{\link{as.data.table.dynamitefit}()},
\code{\link{as_draws_df.dynamitefit}()},
\code{\link{coef.dynamitefit}()},
\code{\link{confint.dynamitefit}()},
\code{\link{dynamite}()},
\code{\link{get_code}()},
\code{\link{get_data}()},
\code{\link{get_parameter_dims}()},
\code{\link{get_parameter_names}()},
\code{\link{get_parameter_types}()},
\code{\link{ndraws.dynamitefit}()}
}
\concept{output}
