% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dynamitefit}
\alias{plot.dynamitefit}
\alias{plot_deltas}
\alias{plot_betas}
\alias{plot_nus}
\title{Plot Regression Coefficients of a Dynamite Model}
\usage{
\method{plot}{dynamitefit}(x, ...)

plot_deltas(
  x,
  level = 0.05,
  alpha = 0.5,
  scales = c("fixed", "free"),
  include_alpha = TRUE
)

plot_betas(x, level = 0.05, include_alpha = TRUE)

plot_nus(x, level = 0.05)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object}

\item{...}{Ignored.}

\item{level}{[\code{numeric(1)}]\cr Level for posterior intervals.
Default is 0.05, leading to 90\% intervals.}

\item{alpha}{[\code{numeric(1)}]\cr Opacity level for \code{geom_ribbon}.
Default is 0.5.}

\item{scales}{[\code{character(1)}] Should y-axis of the panels be \code{"fixed"}
(the default) or \code{"free"}? See \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{include_alpha}{[\code{logical(1)}]\cr If \code{TRUE} (default), extracts also
time-invariant intercept term alpha if time-invariant parameters beta are
extracted, and time-varying alpha if time-varying delta are extracted.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Regression Coefficients of a Dynamite Model
}
\section{Functions}{
\itemize{
\item \code{plot_deltas}: Visualize Time-varying Regression Coefficients of a Dynamite Model

\item \code{plot_betas}: Time-invariant Regression Coefficients of a Dynamite Model

\item \code{plot_nus}: Visualize Random Intercepts of a Dynamite Model
}}

\examples{
plot(gaussian_example_fit)
plot_deltas(gaussian_example_fit, scales = "free") +
  ggplot2::theme_minimal()
plot_betas(gaussian_example_fit)
plot_nus(gaussian_example_fit)

}
