% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{plot_deltas}
\alias{plot_deltas}
\title{Visualize Time-varying Regression Coefficients of the Dynamite Model}
\usage{
plot_deltas(model, level = 0.05, alpha = 0.5, scales = "fixed")
}
\arguments{
\item{model}{An object of class \code{dynamitefit}.}

\item{level}{[\code{numeric(1)}]\cr Level for posterior intervals.
Default is 0.05, leading to 90\% intervals.}

\item{alpha}{[\code{numeric(1)}]\cr Opacity level for \code{geom_ribbon}.
Default is 0.5.}

\item{scales}{[\code{character(1)}] Should y-axis of the panels be \code{"fixed"}
(the default) or \code{"free"}? See \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize Time-varying Regression Coefficients of the Dynamite Model
}
\examples{
plot_deltas(gaussian_example_fit, scales = "free") +
  ggplot2::theme_minimal()
}
