% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nus}
\alias{plot_nus}
\title{Plot Random Intercepts of a Dynamite Model}
\usage{
plot_nus(x, responses = NULL, level = 0.05)
}
\arguments{
\item{x}{[\code{dynamitefit}]\cr The model fit object}

\item{responses}{[\code{character()}]\cr Response(s) for which the coefficients
should be drawn. Possible options are elements of
\code{unique(x$priors$response)}, and the default is this whole vector.}

\item{level}{[\code{numeric(1)}]\cr Level for posterior intervals.
Default is 0.05, leading to 90\% intervals.}
}
\description{
Plot Random Intercepts of a Dynamite Model
}
\examples{
plot_nus(gaussian_example_fit)

}
