% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nus}
\alias{plot_nus}
\title{Visualize Random Intercepts of the Dynamite Model}
\usage{
plot_nus(model, level = 0.05)
}
\arguments{
\item{model}{An object of class \code{dynamitefit}.}

\item{level}{[\code{numeric(1)}]\cr Level for posterior intervals.
Default is 0.05, leading to 90\% intervals.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize Random Intercepts of the Dynamite Model
}
\examples{
plot_nus(gaussian_example_fit)
}
