% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.dynamitefit}
\alias{predict.dynamitefit}
\title{Predict method for a Bayesian Time-Varying Coefficients Model}
\usage{
\method{predict}{dynamitefit}(
  object,
  newdata = NULL,
  type = c("response", "mean", "link"),
  impute = c("none", "locf", "linear"),
  n_draws = NULL,
  ...
)
}
\arguments{
\item{object}{[\code{dynamitefit}]\cr The model fit object.}

\item{newdata}{[\code{data.frame}]\cr Data used in predictions.
If \code{NULL} (default), the data used in model estimation is used for
predictions as well.}

\item{type}{[\code{character(1)}]\cr Type of prediction,
\code{"response"} (default), \code{"mean"}  or \code{"link"}.}

\item{impute}{[\code{character(1)}]\cr Which imputation scheme to use for
missing predictor values. Currently supported options are
no imputation: \code{"none"} (default), and
last observation carried forward: \code{"locf"}.}

\item{n_draws}{[\code{integer(1)}]\cr Number of posterior samples to use,
default is \code{NULL} which uses all samples.}

\item{...}{Ignored.}
}
\description{
Predict method for a Bayesian Time-Varying Coefficients Model
}
