% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.dynamitefit}
\alias{predict.dynamitefit}
\title{Predict method for a Bayesian Time-Varying Coefficients Model}
\usage{
\method{predict}{dynamitefit}(
  object,
  newdata = NULL,
  mode = c("counterfactual", "forecast"),
  type = c("response", "mean", "link"),
  n_draws = NULL
)
}
\arguments{
\item{object}{Object of class \code{dynamitefit}.}

\item{newdata}{Data frame used in predictions. If \code{NULL} (default), the
data used in model estimation is used for predictions as well.}

\item{mode}{TODO: Think about the names, if we predict missing observations
does the term counterfactual make sense?}

\item{type}{Type of prediction, \code{"response"} (default), \code{"mean"}
or \code{"link"}.}

\item{n_draws}{Number of posterior samples to use, default is all.}
}
\description{
Predict method for a Bayesian Time-Varying Coefficients Model
}
