% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.dynamitefit}
\alias{predict.dynamitefit}
\title{Predict method for a Bayesian Time-Varying Coefficients Model}
\usage{
\method{predict}{dynamitefit}(
  object,
  newdata = NULL,
  type = c("response", "mean", "link"),
  impute = c("none", "locf"),
  new_levels = c("none", "bootstrap", "gaussian", "original"),
  n_draws = NULL,
  ...
)
}
\arguments{
\item{object}{[\code{dynamitefit}]\cr The model fit object.}

\item{newdata}{[\code{data.frame}]\cr Data used in predictions. Predictions are
computed for missing (\code{NA}) values in the response variable columns, and
non-missing values are assumed fixed.
If \code{NULL} (default), the data used in model estimation is used for
predictions as well, after all values in the response variable columns
after the first \code{fixed} time points are converted to \code{NA} values.
Missing values in predictor columns can be imputed (argument \code{impute}).
There should be no new time points that were not present in the data that
were used to fit the model. New group levels can be included, but if the
model contains random intercepts, an options for the random
effects for the new levels must be chosen (argument \code{new_levels}). Note
that as \code{newdata} is expanded with predictions, it can be beneficial
in terms of memory usage to first remove redundant columns from \code{newdata}.}

\item{type}{[\code{character(1)}]\cr Type of prediction,
\code{"response"} (default), \code{"mean"}, or \code{"link"}.}

\item{impute}{[\code{character(1)}]\cr Which imputation scheme to use for
missing predictor values. Currently supported options are
no imputation: \code{"none"} (default), and
last observation carried forward: \code{"locf"}.}

\item{new_levels}{[\code{character(1)}: \sQuote{NULL}]\cr
Defines if and how to sample the random intercepts for observations whose
group level was not present in the original data. The options are
* \code{"none"} (the default) which will signal an error if new levels
are encountered.
* \code{"bootstrap"} which will randomly draw from the posterior samples of
the random intercepts across all original levels.
* \code{"gaussian"} which will randomly draw from a gaussian
distribution using the posterior samples of the random intercept
standard deviation.
* \code{"original"} which will randomly match each new level to one of
the original levels. The posterior samples of the random intercept of
the matched levels will then be used for the new levels.
This argument is ignored if model does not contain random intercepts.}

\item{n_draws}{[\code{integer(1)}]\cr Number of posterior samples to use,
default is \code{NULL} which uses all samples.}

\item{...}{Ignored.}
}
\description{
Predict method for a Bayesian Time-Varying Coefficients Model
}
\examples{
predict(gaussian_example_fit, type = "response", n_draws = 2L)
}
