% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.dynamitefit}
\alias{predict.dynamitefit}
\title{Predict Method for a Dynamite Model}
\usage{
\method{predict}{dynamitefit}(
  object,
  newdata = NULL,
  type = c("response", "mean", "link"),
  funs = list(),
  impute = c("none", "locf"),
  new_levels = c("none", "bootstrap", "gaussian", "original"),
  global_fixed = FALSE,
  n_draws = NULL,
  expand = TRUE,
  ...
)
}
\arguments{
\item{object}{[\code{dynamitefit}]\cr The model fit object.}

\item{newdata}{[\code{data.frame}]\cr Data used in predictions. Predictions are
computed for missing (\code{NA}) values in the response variable columns, and
non-missing values are assumed fixed.
If \code{NULL} (default), the data used in model estimation is used for
predictions as well, after all values in the response variable columns
after the first \code{fixed} time points are converted to \code{NA} values.
Missing values in predictor columns can be imputed (argument \code{impute}).
There should be no new time points that were not present in the data that
were used to fit the model. New group levels can be included, but if the
model contains random intercepts, an options for the random
effects for the new levels must be chosen (argument \code{new_levels}). Note
that as \code{newdata} is expanded with predictions, it can be beneficial
in terms of memory usage to first remove redundant columns from \code{newdata}.}

\item{type}{[\code{character(1)}]\cr Type of prediction,
\code{"response"} (default), \code{"mean"}, or \code{"link"}.}

\item{funs}{[\code{list()}]\cr A named list whose names should correspond to the
response variables of the model. Each element of \code{funs} should be a
a named \code{list} of functions that will be applied to the
corresponding predicted responses of the channel over the individuals
for each combination of the posterior draws and time points.
In other words, the resulting predictions will be averages
over the individuals. The functions should take the corresponding
response variable values as their only argument.
If \code{funs} is empty, the full individual level predictions are returned
instead. Note that this argument can only be used
if there are multiple individuals (i.e., \code{group} was not \code{NULL} in the
\code{dynamite} call).}

\item{impute}{[\code{character(1)}]\cr Which imputation scheme to use for
missing predictor values. Currently supported options are
no imputation: \code{"none"} (default), and
last observation carried forward: \code{"locf"}.}

\item{new_levels}{[\code{character(1)}]\cr
Defines if and how to sample the random intercepts for observations whose
group level was not present in the original data. The options are
* \code{"none"} (the default) which will signal an error if new levels
are encountered.
* \code{"bootstrap"} which will randomly draw from the posterior samples of
the random intercepts across all original levels.
* \code{"gaussian"} which will randomly draw from a gaussian
distribution using the posterior samples of the random intercept
standard deviation.
* \code{"original"} which will randomly match each new level to one of
the original levels. The posterior samples of the random intercept of
the matched levels will then be used for the new levels.
This argument is ignored if model does not contain random intercepts.}

\item{global_fixed}{[\code{logical(1)}]\cr If \code{FALSE} (the default),
the first non-fixed time point is counted from the the first non-NA
observation for each group member separately. Otherwise, the first
non-fixed time point is counted from the first time point globally.
If there are no groups, then the options are equivalent.}

\item{n_draws}{[\code{integer(1)}]\cr Number of posterior samples to use,
default is \code{NULL} which uses all samples.}

\item{expand}{[\code{logical(1)}]\cr If \code{TRUE} (the default), the output
is a single \code{data.frame} containing the original \code{newdata} and the
predicted values. Otherwise, a \code{list} is returned with two components,
\code{simulated} and \code{observed}, where the first contains only the
predicted values, and the second contains the original \code{newdata}.
Setting \code{expand} to \code{FALSE} can help conserve memory because \code{newdata}
is not replicated \code{n_draws} times in the output.
This argument is ignored if \code{funs} are provided.}

\item{...}{Ignored.}
}
\value{
A \code{data.frame} containing the predicted values or a \code{list} of two
\code{data.frames}. See the \code{expand} argument for details.
}
\description{
Obtain counterfactual predictions for a \code{dynamitefit} object. Note that
forecasting (i.e., predictions for time indices beyond the last time index
in the original data) is not supported by the \pkg{dynamite} package.
}
\examples{
predict(gaussian_example_fit, type = "response", n_draws = 2L)

}
