% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dynamitefit}
\alias{print.dynamitefit}
\title{Print a Summary of a Dynamite Object}
\usage{
\method{print}{dynamitefit}(x, ...)
}
\arguments{
\item{x}{[\code{dynamitefit}] The model fit object.}

\item{...}{Further arguments to the print method for tibbles.
See \link[tibble:formatting]{tibble::formatting}.}
}
\description{
Prints the summary information of the estimated dynamite model: The smallest
effective sample sizes, largest Rhat and summary statistics of the
time-invariant model parameters.
}
\examples{
print(gaussian_example_fit)
}
