% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo.R
\name{print.lfo}
\alias{print.lfo}
\title{Print the results from the LFO}
\usage{
\method{print}{lfo}(x, ...)
}
\arguments{
\item{x}{[\code{lfo}]\cr Output of the \code{lfo} method.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Prints the summary of the leave-future-out cross-validation.
}
\examples{
data.table::setDTthreads(1) # For CRAN
\donttest{
# Please update your rstan and StanHeaders installation before running
# on Windows
if (!identical(.Platform$OS.type, "windows")) {
  # This gives warnings due to the small number of iterations
  suppressWarnings(lfo(gaussian_example_fit, L = 20))
}
}

}
