% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random}
\alias{random}
\title{Define the Random Intercepts for the Dynamite Model.}
\usage{
random(channels = NULL, correlated = TRUE)
}
\arguments{
\item{channels}{[\code{character()}]\cr Names of the channels for which the
random intercepts should be defined. Default is all channels.}

\item{correlated}{[\code{logical(1)}]\cr If \code{TRUE} (the default), correlation of
intercepts within a group (i.e. between channels) is modelled
(as multivariate normal).}
}
\value{
An object of class \code{random}.
}
\description{
This function can be used as part of \code{dynamiteformula} to define random
intercepts for each group.
}
