% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random}
\alias{random}
\title{Define Random Intercepts for the Dynamite Model.}
\usage{
random(responses = NULL, correlated = TRUE)
}
\arguments{
\item{responses}{[\code{character()}]\cr Names of the responses for which the
random intercepts should be defined. Default is all responses defined with
\code{obs}.}

\item{correlated}{[\code{logical(1)}]\cr If \code{TRUE} (the default), correlations of
intercepts within a group (i.e., between responses) are modeled so that
the intercepts follow a multivariate normal distribution.}
}
\value{
An object of class \code{random}.
}
\description{
This function can be used as part of \code{\link[=dynamiteformula]{dynamiteformula()}} to define
(correlated) random intercepts for each group.
}
\examples{
# three channel model with correlated random effects for responses x and y
obs(y ~ 1, family = "gaussian") +
  obs(x ~ 1, family = "poisson") +
  obs(z ~ 1, family = "gaussian") +
  random(responses = c("y", "x"), correlated = TRUE)

}
