% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{splines}
\alias{splines}
\title{Define the B-splines Used for the Time-varying Coefficients of the Model.}
\usage{
splines(
  shrinkage = FALSE,
  override = FALSE,
  df = NULL,
  degree = 3,
  lb_tau = 0,
  noncentered = FALSE
)
}
\arguments{
\item{shrinkage}{[\code{logical(1)}]\cr If \code{TRUE}, a common global shrinkage
parameter \eqn{\lambda} is used for the splines so that the standard
deviation of the random walk prior is of the spline coefficients is
\eqn{\lambda\tau}. Default is \code{FALSE}.
TODO: Needs more testing whether this is useful.}

\item{override}{[\code{logical(1)}]\cr If \code{FALSE} (the default), an existing
definition for the splines will not be overridden by another call to
\code{splines()}. If \code{TRUE}, any existing definitions will be replaced.}

\item{df}{[\code{numeric()}]\cr Degree of freedom, i.e. the total number of
spline coefficients. See \code{\link[splines:bs]{splines::bs()}}. Note that the knots are always
defined as equidistant sequence on the interval starting from the first
non-fixed time point to the last time point in the data.}

\item{degree}{[\code{integer(1)}]\cr See \code{\link[splines:bs]{splines::bs()}}.}

\item{lb_tau}{[\code{numeric()}]\cr Hard constraint(s) on the lower bound of the
standard deviation parameters \eqn{\tau} of the random walk priors. Can be
useful in avoiding divergences in some cases. See also \code{noncentered}
argument. Can be a single positive value, or vector defining the
lower bound separately for each channel.}

\item{noncentered}{[\code{logical()}]\cr If \code{TRUE}, use noncentered
parameterization for the spline coefficients. Default is \code{FALSE}. Try
changing this if you encounter divergences or other problems in sampling.
Can be a single logical value, or vector of logical values, defining the
parameterization separately for each channel.}
}
\description{
This function can be used as part of \code{dynamiteformula} to define the splines
for the time-varying coeffients \eqn{\delta}.
}
\details{
TODO: Think the naming of this option, we have a "local" shrinkage due to
the random walk prior of the spline coefficients, and the this "global" shrinkage over
splines. Need to avoid confusion with the traditional global-local shrinkage priories (e.g. horseshoe)
}
