% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{splines}
\alias{splines}
\title{Define the B-splines used for the regression coefficients of the model.}
\usage{
splines(
  shrinkage = FALSE,
  override = FALSE,
  df = NULL,
  knots = NULL,
  degree = 3,
  Boundary.knots = NULL,
  lb_tau = 0,
  noncentered = FALSE
)
}
\arguments{
\item{shrinkage}{a logical value. If `TRUE`` (the default) a shrinkage prior is used
for smoothing the splines. Default is FALSE. TODO: Needs more testing.}

\item{override}{a logical value. If \code{FALSE} (the default), an existing
definition for the splines will not be overridden by another call to \code{splines}.
If \code{TRUE}, any existing definitions will be replaced.}

\item{df}{see \code{\link[splines:bs]{splines::bs()}}.}

\item{knots}{see \code{\link[splines:bs]{splines::bs()}}.}

\item{degree}{see \code{\link[splines:bs]{splines::bs()}}.}

\item{Boundary.knots}{see \code{\link[splines:bs]{splines::bs()}}.}

\item{lb_tau}{TODO.}

\item{noncentered}{TODO.}
}
\description{
TODO: Think the naming of this option, we have a "local" shrinkage due to
the random walk prior of the spline coefficients, and the this "global" shrinkage over
splines. Need to avoid confusion with the traditional global-local shrinkage priories (e.g. horseshoe)
}
