% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{splines}
\alias{splines}
\title{Define the B-splines used for the regression coefficients of the model.}
\usage{
splines(shrinkage = TRUE, override = FALSE)
}
\arguments{
\item{shrinkage}{A logical value. If \code{TRUE} (the default) a shrinkage prior is used
for smoothing the splines.}

\item{override}{A logical value. If \code{FALSE} (the default), an existing
definition for the splines will not be overridden by another call to \code{splines}.
If \code{TRUE}, any existing definitions will be replaced.}
}
\description{
Define the B-splines used for the regression coefficients of the model.
}
