% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dynamitefit}
\alias{summary.dynamitefit}
\title{Summary for a Dynamite Model Fit}
\usage{
\method{summary}{dynamitefit}(
  x,
  row.names = NULL,
  optional = FALSE,
  responses = NULL,
  types = NULL,
  summary = TRUE,
  probs = c(0.05, 0.95),
  include_fixed = TRUE,
  ...
)
}
\arguments{
\item{...}{Further arguments to the \code{as.data.frame.dynamitefit}.}

\item{object}{[\code{dynamitefit}]\cr The model fit object.}
}
\description{
Returns the summary statistics of the posterior samples of the model; this
is an alias of \code{as.data.frame(object)}.
}
\examples{
summary(gaussian_example_fit, types = "beta")

}
