% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dynamitefit}
\alias{summary.dynamitefit}
\title{Summary for a Dynamite Model Fit}
\usage{
\method{summary}{dynamitefit}(object, ...)
}
\arguments{
\item{object}{[\code{dynamitefit}]\cr The model fit object.}

\item{...}{TODO}
}
\description{
Summary for a Dynamite Model Fit
}
\examples{
summary(gaussian_example_fit)

}
